refactor : Member Exception처리
