+++ src/main/java/com/buildup/nextQuestion/dto/solving/FindQuestionByNormalExamRequest.java
@@ -1,8 +1,9 @@
-package com.buildup.nextQuestion.dto.question;
+package com.buildup.nextQuestion.dto.solving;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import com.buildup.nextQuestion.dto.question.*;
 
 @Data
 @AllArgsConstructor
